Set up project structure

+++ src/api/router.py
@@ -1,48 +1,15 @@
-from typing import Annotated
-
-import uvicorn
-from fastapi import Depends, FastAPI, HTTPException
-from pydantic import BaseModel
+from fastapi import APIRouter, HTTPException
 from sqlalchemy import select
-from sqlalchemy.ext.asyncio import (AsyncSession, async_sessionmaker,
-                                    create_async_engine)
-from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
-app = FastAPI()
+from src.database import Base, engine
+from src.api.dependencies import SessionDepends
+from src.models import NoteModel
+from src.schemas import NoteAddSchema
 
-engine = create_async_engine('sqlite+aiosqlite:///notes.db', echo=True)
-new_session = async_sessionmaker(engine, expire_on_commit=False)
-
-
-async def get_session():
-    async with new_session() as session:
-        yield session
+router = APIRouter()
 
 
-SessionDepends = Annotated[AsyncSession, Depends(get_session)]
-
-
-class NoteAddSchema(BaseModel):
-    header: str
-    description: str
-
-
-class Note(NoteAddSchema):
-    id: int
-
-
-class Base(DeclarativeBase):
-    pass
-
-
-class NoteModel(Base):
-    __tablename__ = 'notes'
-    id: Mapped[int] = mapped_column(primary_key=True)
-    header: Mapped[str]
-    description: Mapped[str]
-
-
-@app.post('/initialization')
+@router.post('/initialization')
 async def init_db():
     async with engine.begin() as conn:
         await conn.run_sync(Base.metadata.drop_all)
@@ -50,7 +17,7 @@ async def init_db():
     return {'configured': True}
 
 
-@app.post('/new_note')
+@router.post('/new_note')
 async def add_note(data: NoteAddSchema, session: SessionDepends):
     new_note = NoteModel(
         header=data.header,
@@ -61,14 +28,14 @@ async def add_note(data: NoteAddSchema, session: SessionDepends):
     return {"note added": True}
 
 
-@app.get('/all_notes')
+@router.get('/all_notes')
 async def get_all_notes(session: SessionDepends):
     query = select(NoteModel)
     result = await session.execute(query)
     return result.scalars().all()
 
 
-@app.get('/notes/{note_id}')
+@router.get('/notes/{note_id}')
 async def get_note(note_id: int, session: SessionDepends):
     query = select(NoteModel).filter_by(id=note_id)
     result = await session.execute(query)
@@ -78,7 +45,7 @@ async def get_note(note_id: int, session: SessionDepends):
     return note
 
 
-@app.put('/notes/{note_id}')
+@router.put('/notes/{note_id}')
 async def update_note(note_id: int, session: SessionDepends, data: NoteAddSchema):
     query = select(NoteModel).filter_by(id=note_id)
     result = await session.execute(query)
@@ -95,17 +62,15 @@ async def update_note(note_id: int, session: SessionDepends, data: NoteAddSchema
     return note
 
 
-@app.delete("/notes/{note_id}")
-async def delete_not(note_id: int, session: SessionDepends):
+@router.delete("/notes/{note_id}")
+async def delete_note(note_id: int, session: SessionDepends):
     query = select(NoteModel).filter_by(id=note_id)
     result = await session.execute(query)
     note = result.scalar()
+    if note is None:
+        raise HTTPException(status_code=404, detail='this note not found')
 
     await session.delete(note)
     await session.commit()
 
     return {'deleted': True}
-
-
-if __name__ == '__main__':
-    uvicorn.run("main:app", reload=True)
